% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csmtools.R
\docType{package}
\name{csmtools-package}
\alias{csmtools}
\alias{csmtools-package}
\title{csmtools: A collection of useful code}
\description{
csmtools is a collection of code that can be used in different projects to
help prevent team members from reinventing the wheel for some things that
others have already figured out.
}
\details{
Bug reports, additions and enhancement requests are welcomed at:
\url{https://github.com/McClellandLegge/csmtools}
}

