% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreads.R
\name{dread}
\alias{dread}
\title{Fread all files in a directory}
\usage{
dread(dir_path, ...)
}
\arguments{
\item{dir_path}{A valid directory path}

\item{...}{Additional arguments to \code{\link[data.table]{fread}}}
}
\value{
A data table
}
\description{
Fread all files in a directory
}

