% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_reduce.R
\name{dt_reduce}
\alias{dt_reduce}
\title{Apply a row-wise Reduce}
\usage{
dt_reduce(DT, FUN, ...)
}
\arguments{
\item{DT}{A data.table}

\item{FUN}{Any binary function}

\item{...}{Quoted column names from DT}
}
\value{
A vector

A vector, class will vary
}
\description{
Apply a row-wise Reduce
}
\details{
Apply a row-wise reduce for a given function on a set of a data.table's columns
}
\examples{
library("data.table")
DT <- as.data.table(head(iris))
# basic differencing
dt_reduce(DT, `-`, "Sepal.Length", "Sepal.Width")

# paste columns together row-wise
dt_reduce(DT, paste, colnames(DT))

# calculate the mean
dt_reduce(DT, `+`, "Sepal.Length", "Sepal.Width", "Petal.Length") / nrow(DT)
}

