% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_empty_files.R
\name{filter_empty_files}
\alias{filter_empty_files}
\title{Filter out file names with non-zero rowcounts}
\usage{
filter_empty_files(x)
}
\arguments{
\item{x}{A character vector of filenames}
}
\value{
A character vector
}
\description{
Filter out file names with non-zero rowcounts
}

