% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_color_hue.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Emulate the default ggplot2 color palette}
\usage{
gg_color_hue(n, rgb = TRUE)
}
\arguments{
\item{n}{The number of colors to produce}

\item{rgb}{A Boolean, should a \code{rgb(<r>, <g>, <b>)} character string
be returned instead of the hex values?}
}
\value{
A character vector
}
\description{
Emulate the default ggplot2 color palette
}
\details{
An adapted function from John Colby's \link[answer]{http://stackoverflow.com/a/8197703/3034614}
   on how to emulate the ggplot2 default color palette, which is just equal
   spacing on the color wheel.
}
\examples{
gg_color_hue(5)
gg_color_hue(5, rgb = FALSE)
}

