% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_hive.R
\name{has_hive}
\alias{has_hive}
\title{Check if the system has hive capabilities}
\usage{
has_hive()
}
\value{
A boolean
}
\description{
Check if the system has hive capabilities
}
\details{
Checks to see if the hive binaries are in the PATH variable
}
\examples{
if(has_hive()) {
    print("yes")
} else {
    print("no")
}
}

