% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hive_datatypes.R
\name{hive_datatypes}
\alias{hive_datatypes}
\title{Extract the R-datatypes for a hive table}
\usage{
hive_datatypes(schema, table_name)
}
\arguments{
\item{schema}{A character string, the name of the hive schema}

\item{table_name}{A character string, the name of the hive table}
}
\value{
A data.table
}
\description{
Extract the R-datatypes for a hive table
}

