% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iri_week_date.R
\name{iri_week_date}
\alias{iri_week_date}
\title{A function to derive the week-ending or week-begining date for an IRI week}
\usage{
iri_week_date(x, week_ending = TRUE)
}
\arguments{
\item{x}{An IRI week}

\item{week_ending}{A boolean, should we return the date for week-ending (\code{TRUE})
or week-beginning (\code{FALSE})}
}
\value{
A Date
}
\description{
A function to derive the week-ending or week-begining date for an IRI week
}
\examples{
iri_week_date(1900, week_ending = FALSE)
}
\seealso{
\link{}
}

