% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_defined.R
\name{is_defined}
\alias{is_defined}
\title{A function to perform a read of a hive table}
\usage{
is_defined(..., .all = FALSE)
}
\arguments{
\item{...}{A list of objects to test if they are null}

\item{.all}{Boolean, should we return the tests for each individual element?}
}
\description{
A function to perform a read of a hive table
}
\details{
Good for testing an input(s) to a function when it might be NULL
}
\examples{
foo <- function(x = NULL, y = NULL) {
  if (is_defined(x, y)) {
    return(paste0(x, y))
  } else if (is_defined(x)) {
    return(x)
  } else if (any(is_defined(x, y, .all = TRUE))) {
    return("one is not null")
  }
}
foo(x = 1)
foo(x = 1, y = 2)
foo(y = 2)
}

