% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_color.R
\name{make_color}
\alias{make_color}
\title{A function to convert colors to their hex values}
\usage{
make_color(...)
}
\arguments{
\item{...}{A (possibly) mixed-type vector}
}
\value{
A list of hex values or Boolean \code{FALSE} when the element cannot
   be interpreted as a color
}
\description{
A function to convert colors to their hex values
}
\details{
A list is returned instead of a vector to avoid the coercion of a
   Boolean value to a character one
}
\examples{
make_color(NA, "black", "blackk", 5, "#00", "#000000", "rgb(1, 1, 1, 0.5)")
}

