% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upc.R
\name{make_upc}
\alias{make_upc}
\title{Make a upc form component parts}
\usage{
make_upc(sys, gen, ven, ite, format = "sgvi")
}
\arguments{
\item{sys}{A vector that can be coerced to integer}

\item{gen}{A vector that can be coerced to integer}

\item{ven}{A vector that can be coerced to integer}

\item{ite}{A vector that can be coerced to integer}

\item{format}{A character string with the four letters "s", "g", "v", and "i"
arranged in the desired format of the outout, e.g. 'svig'}
}
\value{
A character vector
}
\description{
Make a upc form component parts
}
\examples{
set.seed(1)
sys <- rep(0, 5)
gen <- rep(1, 5)
ven <- sample(10000:99999, 5)
ite <- sample(10000:99999, 5)
make_upc(sys, gen, ven, ite, format = "sgiv")
}

