% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mread.R
\name{mread}
\alias{mread}
\title{"Meta-read": Read a file with an accompanying meta file}
\usage{
mread(filepath, metafilepath = NULL, ...)
}
\arguments{
\item{filepath}{The file path of the file to be read}

\item{metafilepath}{Default NULL, the file path of the meta file. Assumes the
file is in the same directory and has an identical naming convention}

\item{...}{Other arguments to \code{\link[data.table]{fread}}}
}
\value{
A data.table object
}
\description{
"Meta-read": Read a file with an accompanying meta file
}
\details{
Silently assigns the meta table to the global environment
}

