% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paman.R
\name{paman}
\alias{paman}
\title{Element-wise paste of a matrix using the column names}
\usage{
paman(x, ...)
}
\arguments{
\item{x}{A matrix, or object able to be coerced to a matric with non-null
dimnames}

\item{...}{Arguments to paste}
}
\value{
A matrix
}
\description{
Element-wise paste of a matrix using the column names
}
\details{
Works just like a normal \link[base]{paste0} function except the input
   is expected to be a matrix and the output will likewise be a matrix. Its
   a "paste" for a "matrix" "element" = "pa" + "m" + "el" = "pamel". Helpful
   for parsing hovertext for 3D plotly objects.
}
\examples{
m <- matrix(runif(9), nrow = 3, dimnames = list(LETTERS[1:3], LETTERS[4:6]))

paman(m, "From: ", y, " to: ", x)
paman(m, "From: ", x, " to: ", y)
}
\seealso{
\link{pamat} \link{pamel}
}

