% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamat.R
\name{pamat}
\alias{pamat}
\title{Element-wise paste of two (or more) matrices with a separator}
\usage{
pamat(..., sep = " ")
}
\arguments{
\item{...}{Matrix objects}

\item{sep}{A character string, the separator between the element-wise paste}
}
\value{
A matrix
}
\description{
Element-wise paste of two (or more) matrices with a separator
}
\details{
The matrices will be pasted together in the order in which they are
    specified and one separator will be shared used. Its a "paste" for a
    "matrix" "element" = "pa" + "m" + "el" = "pamel". Helpful for parsing
    hovertext for 3D plotly objects.
}
\examples{
caps <- matrix(LETTERS[1:9], nrow = 3)
lows <- matrix(letters[1:9], nrow = 3)
pamat(caps, lows, sep = " -> ")
}
\seealso{
\link{pamel} \link{paman}
}

