% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_palette.R
\name{preview_palatte}
\alias{preview_palatte}
\title{Preview your hex color palatte}
\usage{
preview_palatte(x)
}
\arguments{
\item{x}{A vector of hex colors}
}
\description{
Preview your hex color palatte
}
\details{
Plots a simple image with swaths of the colors specified in the
   palatte, in the order in which they are specified
}
\examples{
colfunc <- colorRampPalette(c("white", "dodgerblue"))
my_cols <- colfunc(20)
preview_palatte(my_cols)
}

