% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_.R
\name{seq_}
\alias{seq_}
\title{Extension of base::seq to handle range specification}
\usage{
seq_(range, ...)
}
\arguments{
\item{range}{A numeric vector of length two specifying the 'from' and 'to'
arguments to the \code{\link[base]{seq}} function}

\item{...}{Other arguments passed to \code{\link[base]{seq}}}
}
\value{
A vector
}
\description{
Extension of base::seq to handle range specification
}
\examples{
seq_(range(1:10), by = 2)
}

