% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upc.R
\name{split_upc}
\alias{split_upc}
\title{Split up a character UPC}
\usage{
split_upc(upc, format = "sgvi", convert = FALSE)
}
\arguments{
\item{upc}{A character vector}

\item{format}{A character string with the four letters "s", "g", "v", and "i"
arranged in the format of the input upc, e.g. 'svig'}

\item{convert}{A boolean, should the final component pieces be converted to
numeric?}
}
\value{
A named list of the component pieces
}
\description{
Split up a character UPC
}
\examples{
upcs <- c("0000790350211", "0007403067401", "0010248198421")
split_upc(upcs, format = "svig", convert = TRUE)
}

