% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_compare.R
\name{dt_compare}
\alias{dt_compare}
\title{Merge and compare columns of data.frames (data.tables)}
\usage{
dt_compare(x, y, compare = NULL, func = `-`, round = TRUE,
  precision = 6, verbose = TRUE, plot = TRUE, ...)
}
\arguments{
\item{x}{A data.frame}

\item{y}{A data.frame}

\item{compare}{A character string or vector of shared column names}

\item{func}{A binary function to compare the columns with, should be appropriate
for the datatypes of the columns}

\item{round}{A boolean, should we round at all?}

\item{precision}{The precision of the comparison, is the \code{digits}
argument to the \code{\link[base]{round}} function}

\item{verbose}{A boolean, should messages be written to the window?}

\item{plot}{A boolean, should a summary plot be produced? Requires the loading
of additional packages}

\item{...}{Any arguments to the \code{\link[data.table]{merge}} function}
}
\value{
A data.frame
}
\description{
Merge and compare columns of data.frames (data.tables)
}
\examples{
x <- iris[1:50,]
y <- iris[1:60,]
x$id <- seq(nrow(x))
y$id <- seq(nrow(y))
y$Sepal.Width = y$Sepal.Width + rnorm(n = nrow(y))

# can specify any arguments to 'merge'
res <- dt_compare(x, y, compare = c("Sepal.Width", "Sepal.Length"), by = "id", all.y = TRUE)
}

