% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_size_filter.R
\name{file_size_filter}
\alias{file_size_filter}
\title{Filter files based on their size}
\usage{
file_size_filter(x, size = 0, units = "B", include = FALSE)
}
\arguments{
\item{x}{A character vector of filenames}

\item{size}{A numeric vector, the size of the file}

\item{units}{A character vector specifying the units. Options are \code{B}
, \code{KB}, \code{MB} and \code{GB}. Must match the length of \code{size}
if specifying more than one unit.}

\item{include}{A boolean, include files of size \code{size}?}
}
\value{
A character vector
}
\description{
Filter files based on their size
}
\details{
Filters out the files that are less than (or less than or equal to)
   the size with units specified. If all files are filtered out, then a
   character vector of length 0 is returned.
}
\examples{
x <- list.files(path = Sys.getenv("TEMP"), full.names = TRUE)[1]
file_size_filter(x, size = 1, units = "KB")
}

