% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_files.R
\name{filter_files}
\alias{filter_files}
\title{Filter out file names based on criteria}
\usage{
filter_files(x, size = 0, units = "B", include = FALSE, simplify = TRUE)
}
\arguments{
\item{x}{A character vector of filenames}

\item{size}{A numeric vector, the size of the file}

\item{units}{A character vector specifying the units. Options are \code{B}
, \code{KB}, \code{MB} and \code{GB}. Must match the length of \code{size}
if specifying more than one unit.}

\item{include}{A boolean, include files of size \code{size}?}

\item{simplify}{A boolean, should we create a data.table from the list of output?}
}
\value{
A character vector
}
\description{
Filter out file names based on criteria
}
\details{
Automatically excludes any non-existant items
}
\examples{

x <- list.files(path = Sys.getenv("TEMP"), full.names = TRUE)[1:100]
filter_files(x, size = 1, units = "KB")
}

