% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hive_read.R
\name{hive_read}
\alias{hive_read}
\title{Read a Hive table that is stored as a text file}
\usage{
hive_read(x, data_types, col_names, ...)
}
\arguments{
\item{x}{A character string, the directory path of the hive table to read}

\item{data_types}{A character vector, the data types of the columns, in order}

\item{col_names}{A character vector, the column names, in order}

\item{...}{Additional arguments to \code{\link[data.table]{fread}}}
}
\value{
A data.table
}
\description{
Read a Hive table that is stored as a text file
}
\details{
The function allows you to assign your own additional arguments to
   fread, but it defaults the separator to pipe ("|") and adds to the
   na.string to recognize the hive default.
}

