% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hread.R
\name{hread}
\alias{hread}
\title{A function to perform a read of a hive table}
\usage{
hread(fp, schema, schema_loc, ...)
}
\arguments{
\item{fp}{The name of the hive table}

\item{schema}{The name of the hive schema}

\item{schema_loc}{The directory path of where the schema is located on the HDFS}
}
\value{
A data.table
}
\description{
A function to perform a read of a hive table
}
\details{
Will automatically read all files under the directory after finding
   the datatypes and column names from the hive metastore
}

