% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hread.R
\name{hread}
\alias{hread}
\title{A function to perform a read of a hive table}
\usage{
hread(table_name, schema, schema_loc, ...)
}
\arguments{
\item{table_name}{A character string, the name of the hive table}

\item{schema}{A character string, the name of the hive schema}

\item{schema_loc}{A character string, the directory path of where the schema
is located on the HDFS}

\item{...}{Additional arguments to \code{\link{hive_read}}}
}
\value{
A data.table
}
\description{
A function to perform a read of a hive table
}
\details{
Will automatically read all files under the directory after finding
   the datatypes and column names from the hive metastore. Note that the
   \code{schema} can have a different physical location instead of being forced
   to have the \code{schema_loc/schema} naming convention.
}
\examples{
\dontrun{
schema_loc <- "/mapr/mapr03r/analytic_users/msmck/csm_synd_hive_schemas/csm_syndicated/"
hread("dictionary", "csm_syndicated", schema_loc)
}
}

