% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iri_week.R
\name{iri_week}
\alias{iri_week}
\title{A function to derive the IRI week for the date specified}
\usage{
iri_week(x, fmt = "\%Y-\%m-\%d", ...)
}
\arguments{
\item{x}{A date or character string}

\item{fmt}{A date format}

\item{...}{Additional arguments to \code{as.Date} including further arguments
to be passed from or to other methods, including \code{format} for
\code{as.character} and \code{as.Date} methods.}
}
\value{
A numeric
}
\description{
A function to derive the IRI week for the date specified
}
\examples{
iri_week(Sys.Date())
iri_week("Dec. 12, 2016", "\%b. \%d, \%Y")
}
\seealso{
\link{iri_week_date}
}

