% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iri_week.R
\name{iri_week}
\alias{iri_week}
\title{A function to extend the functionaliy of the base::floor function}
\usage{
iri_week(x, fmt = "\%Y-\%m-\%d")
}
\arguments{
\item{x}{A numeric}

\item{digits}{The number of digits to the left of the decimal place to round to}
}
\value{
A numeric
}
\description{
A function to extend the functionaliy of the base::floor function
}
\details{
Usually used for purchase data when you need to floor the cents
}
\examples{
x <- 99.9999
floor(x, 2)
# is equivalent to:
base::floor(100 * x)
}

