% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninja_load.R
\name{ninja_load}
\alias{ninja_load}
\title{Load packages silently}
\usage{
ninja_load(...)
}
\arguments{
\item{...}{The quoted names of the packages you wish to load with deadly silence}
}
\description{
Load packages silently
}
\examples{
# load some notoriously loud packages
ninja_load("data.table", "bit64")
}

